Added method InternalServer::RegisterObject() to have the capability to
add annotated business classes to the UA server interface, and provide the
in this class contained, and annotated methods and properties with
UaMethod and UaVariable, respectively, as an interactive server interface,
as well. Please have in mind, that it's really suggested to understand your
UA server interface as an object oriented interface, and program your
business logic that way, too. It's better to understand and better to maintain.

What does it mean?

1. Interact with your system via methods.
2. Represent the internal state system via properties.

That's it!