Can append the events after the InternalServerManager instance is started, only. Did it, and it runs!
I think, so we have the best interface to get the necessary events to the client code. Because of
we don't need delegate the SetConnectCallback()|SetDisconnectCallback() from InternalServer
instance to the InternalServerManager instance to the InternalNodeManager instance, that's
really annoying. And we could replace the OnConnect()|OnDisconnect() methods by lightweight
lambda expressions.