Added the feature to have cascaded UA Object Nodes.
To use it, you've to add the attribute UaObjectList to a ICollection<object> typed property, e. g. List<object>. If one or some of the contained objects have the attribute UaObject, these objects become child UA nodes of the list owner. Thats it.

You can find an example usage on ServerConsole.Program.